refactored some more common components added rich logger
