+++ src/pytorch_tabular/models/common/layers/gated_units.py
@@ -1,0 +1,235 @@
+# Pytorch Tabular
+# Author: Manu Joseph <manujoseph@gmail.com>
+# For license information, see LICENSE.TXT
+import random
+from typing import Callable
+
+import torch
+import torch.nn as nn
+
+from pytorch_tabular.models.common.layers.activations import entmax15, RSoftmax
+
+
+class GatedFeatureLearningUnit(nn.Module):
+    def __init__(
+        self,
+        n_features_in: int,
+        n_stages: int,
+        feature_mask_function: Callable = entmax15,
+        feature_sparsity: float = 0.3,
+        learnable_sparsity: bool = True,
+        dropout: float = 0.0,
+    ):
+        super().__init__()
+        self.n_features_in = n_features_in
+        self.n_features_out = n_features_in
+        self.feature_mask_function = feature_mask_function
+        self._dropout = dropout
+        self.n_stages = n_stages
+        self.feature_sparsity = feature_sparsity
+        self.learnable_sparsity = learnable_sparsity
+        self._build_network()
+
+    def _create_feature_mask(self):
+        feature_masks = torch.cat(
+            [
+                torch.distributions.Beta(
+                    torch.tensor([random.uniform(0.5, 10.0)]),
+                    torch.tensor([random.uniform(0.5, 10.0)]),
+                )
+                .sample((self.n_features_in,))
+                .squeeze(-1)
+                for _ in range(self.n_stages)
+            ]
+        ).reshape(self.n_stages, self.n_features_in)
+        return nn.Parameter(
+            feature_masks,
+            requires_grad=True,
+        )
+
+    def _build_network(self):
+        self.W_in = nn.ModuleList(
+            [
+                nn.Linear(2 * self.n_features_in, 2 * self.n_features_in)
+                for _ in range(self.n_stages)
+            ]
+        )
+        self.W_out = nn.ModuleList(
+            [
+                nn.Linear(2 * self.n_features_in, self.n_features_in)
+                for _ in range(self.n_stages)
+            ]
+        )
+
+        self.feature_masks = self._create_feature_mask()
+        if self.feature_mask_function.__name__ == "t_softmax":
+            t = RSoftmax.calculate_t(
+                self.feature_masks, r=torch.tensor([self.feature_sparsity]), dim=-1
+            )
+            self.t = nn.Parameter(t, requires_grad=self.learnable_sparsity)
+        self.dropout = nn.Dropout(self._dropout)
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        h = x
+        t = (
+            torch.relu(self.t)
+            if self.feature_mask_function.__name__ == "t_softmax"
+            else None
+        )
+        for d in range(self.n_stages):
+            if self.feature_mask_function.__name__ == "t_softmax":
+                feature = self.feature_mask_function(self.feature_masks[d], t[d]) * x
+            else:
+                feature = self.feature_mask_function(self.feature_masks[d]) * x
+            h_in = self.W_in[d](torch.cat([feature, h], dim=-1))
+            z = torch.sigmoid(h_in[:, : self.n_features_in])
+            r = torch.sigmoid(h_in[:, self.n_features_in :])  # noqa: E203
+            h_out = torch.tanh(self.W_out[d](torch.cat([r * h, x], dim=-1)))
+            h = self.dropout((1 - z) * h + z * h_out)
+        return h
+
+
+# https://github.com/labmlai/annotated_deep_learning_paper_implementations/blob/master/labml_nn/transformers/feed_forward.py
+class PositionWiseFeedForward(nn.Module):
+    r"""
+    title: Position-wise Feed-Forward Network (FFN)
+    summary: Documented reusable implementation of the position wise feedforward network.
+
+    # Position-wise Feed-Forward Network (FFN)
+    This is a [PyTorch](https://pytorch.org)  implementation
+    of position-wise feedforward network used in transformer.
+    FFN consists of two fully connected layers.
+    Number of dimensions in the hidden layer $d_{ff}$, is generally set to around
+    four times that of the token embedding $d_{model}$.
+    So it is sometime also called the expand-and-contract network.
+    There is an activation at the hidden layer, which is
+    usually set to ReLU (Rectified Linear Unit) activation, $$\\max(0, x)$$
+    That is, the FFN function is,
+    $$FFN(x, W_1, W_2, b_1, b_2) = \\max(0, x W_1 + b_1) W_2 + b_2$$
+    where $W_1$, $W_2$, $b_1$ and $b_2$ are learnable parameters.
+    Sometimes the
+    GELU (Gaussian Error Linear Unit) activation is also used instead of ReLU.
+    $$x \\Phi(x)$$ where $\\Phi(x) = P(X \\le x), X \\sim \\mathcal{N}(0,1)$
+    ### Gated Linear Units
+    This is a generic implementation that supports different variants including
+    [Gated Linear Units](https://arxiv.org/abs/2002.05202) (GLU).
+    """
+
+    def __init__(
+        self,
+        d_model: int,
+        d_ff: int,
+        dropout: float = 0.1,
+        activation=nn.ReLU(),
+        is_gated: bool = False,
+        bias1: bool = True,
+        bias2: bool = True,
+        bias_gate: bool = True,
+    ):
+        """
+        * `d_model` is the number of features in a token embedding
+        * `d_ff` is the number of features in the hidden layer of the FFN
+        * `dropout` is dropout probability for the hidden layer
+        * `is_gated` specifies whether the hidden layer is gated
+        * `bias1` specified whether the first fully connected layer should have a learnable bias
+        * `bias2` specified whether the second fully connected layer should have a learnable bias
+        * `bias_gate` specified whether the fully connected layer for the gate should have a learnable bias
+        """
+        super().__init__()
+        # Layer one parameterized by weight $W_1$ and bias $b_1$
+        self.layer1 = nn.Linear(d_model, d_ff, bias=bias1)
+        # Layer one parameterized by weight $W_1$ and bias $b_1$
+        self.layer2 = nn.Linear(d_ff, d_model, bias=bias2)
+        # Hidden layer dropout
+        self.dropout = nn.Dropout(dropout)
+        # Activation function $f$
+        self.activation = activation
+        # Whether there is a gate
+        self.is_gated = is_gated
+        if is_gated:
+            # If there is a gate the linear layer to transform inputs to
+            # be multiplied by the gate, parameterized by weight $V$ and bias $c$
+            self.linear_v = nn.Linear(d_model, d_ff, bias=bias_gate)
+
+    def forward(self, x: torch.Tensor):
+        # $f(x W_1 + b_1)$
+        g = self.activation(self.layer1(x))
+        # If gated, $f(x W_1 + b_1) \otimes (x V + b) $
+        if self.is_gated:
+            x = g * self.linear_v(x)
+        # Otherwise
+        else:
+            x = g
+        # Apply dropout
+        x = self.dropout(x)
+        # $(f(x W_1 + b_1) \otimes (x V + b)) W_2 + b_2$ or $f(x W_1 + b_1) W_2 + b_2$
+        # depending on whether it is gated
+        return self.layer2(x)
+
+
+# Inspired by implementations
+# 1. lucidrains - https://github.com/lucidrains/tab-transformer-pytorch/
+# If you are interested in Transformers, you should definitely check out his repositories.
+# 2. PyTorch Wide and Deep - https://github.com/jrzaurin/pytorch-widedeep/
+# It is another library for tabular data, which supports multi modal problems.
+# Check out the library if you haven't already.
+# 3. AutoGluon - https://github.com/awslabs/autogluon
+# AutoGluon is an AuttoML library which supports Tabular data as well. it is from Amazon Research and is in MXNet
+# 4. LabML Annotated Deep Learning Papers - The position-wise FF was shamelessly copied from
+# https://github.com/labmlai/annotated_deep_learning_paper_implementations/tree/master/labml_nn/transformers
+
+
+# GLU Variants Improve Transformer https://arxiv.org/pdf/2002.05202.pdf
+class GEGLU(nn.Module):
+    """Gated Exponential Linear Unit (GEGLU)"""
+
+    def __init__(self, d_model: int, d_ff: int, dropout: float = 0.1):
+        """
+        Args:
+            d_model: dimension of the model
+            d_ff: dimension of the feedforward layer
+            dropout: dropout probability
+        """
+        super().__init__()
+        self.ffn = PositionWiseFeedForward(
+            d_model, d_ff, dropout, nn.GELU(), True, False, False, False
+        )
+
+    def forward(self, x: torch.Tensor):
+        return self.ffn(x)
+
+
+class ReGLU(nn.Module):
+    """ReGLU."""
+
+    def __init__(self, d_model: int, d_ff: int, dropout: float = 0.1):
+        """
+        Args:
+            d_model: dimension of the model
+            d_ff: dimension of the feedforward layer
+            dropout: dropout probability
+        """
+        super().__init__()
+        self.ffn = PositionWiseFeedForward(
+            d_model, d_ff, dropout, nn.ReLU(), True, False, False, False
+        )
+
+    def forward(self, x: torch.Tensor):
+        return self.ffn(x)
+
+
+class SwiGLU(nn.Module):
+    def __init__(self, d_model: int, d_ff: int, dropout: float = 0.1):
+        """
+        Args:
+            d_model: dimension of the model
+            d_ff: dimension of the feedforward layer
+            dropout: dropout probability
+        """
+        super().__init__()
+        self.ffn = PositionWiseFeedForward(
+            d_model, d_ff, dropout, nn.SiLU(), True, False, False, False
+        )
+
+    def forward(self, x: torch.Tensor):
+        return self.ffn(x)
