+++ src/pytorch_tabular/models/common/layers/soft_trees.py
@@ -1,12 +1,15 @@
+import random
+from typing import Callable
 from warnings import warn
 
 import numpy as np
 import torch
 import torch.nn as nn
 
-from .activations import sparsemax, sparsemoid
-from ..layers import ModuleWithInit
 from pytorch_tabular.utils import check_numpy
+
+from ..layers import ModuleWithInit
+from .activations import entmax15, RSoftmax, sparsemax, sparsemoid
 
 # Neural Oblivious Decision Ensembles
 # Author: Sergey Popov, Julian Qian
@@ -196,3 +199,128 @@ class ODST(ModuleWithInit):
             self.tree_dim,
             self.flatten_output,
         )
+
+
+class NeuralDecisionStump(nn.Module):
+    def __init__(
+        self,
+        n_features: int,
+        binning_activation: Callable = entmax15,
+        feature_mask_function: Callable = entmax15,
+        feature_sparsity: float = 0.8,
+        learnable_sparsity: bool = True,
+    ):
+        super().__init__()
+        self._num_cutpoints = 1
+        self._num_leaf = 2
+        self.n_features = n_features
+        self.binning_activation = binning_activation
+        self.feature_mask_function = feature_mask_function
+        self.feature_sparsity = feature_sparsity
+        self.learnable_sparsity = learnable_sparsity
+        self._build_network()
+
+    def _build_network(self):
+        # sampling a random beta distribution
+        # random distribution helps with diversity in trees and feature splits
+        alpha = random.uniform(0.5, 10.0)
+        beta = random.uniform(0.5, 10.0)
+        # with torch.no_grad():
+        feature_mask = (
+            torch.distributions.Beta(torch.tensor([alpha]), torch.tensor([beta]))
+            .sample((self.n_features,))
+            .squeeze(-1)
+        )
+        self.feature_mask = nn.Parameter(feature_mask, requires_grad=True)
+        if self.feature_mask_function.__name__ == "t_softmax":
+            t = RSoftmax.calculate_t(
+                self.feature_mask, r=torch.tensor([self.feature_sparsity])
+            )
+            self.t = nn.Parameter(t, requires_grad=self.learnable_sparsity)
+        W = torch.linspace(
+            1.0,
+            self._num_cutpoints + 1.0,
+            self._num_cutpoints + 1,
+            requires_grad=False,
+        ).reshape(1, 1, -1)
+        self.register_buffer("W", W)
+
+        cutpoints = torch.rand([self.n_features, self._num_cutpoints])
+        # Append zeros to the beginning of each row
+        cutpoints = torch.cat(
+            [torch.zeros([self.n_features, 1], device=cutpoints.device), cutpoints], 1
+        )
+        self.cut_points = nn.Parameter(cutpoints, requires_grad=True)
+        self.leaf_responses = nn.Parameter(
+            torch.rand(self.n_features, self._num_leaf), requires_grad=True
+        )
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        if self.feature_mask_function.__name__ == "t_softmax":
+            t = torch.relu(self.t)
+            feature_mask = self.feature_mask_function(self.feature_mask, t)
+        else:
+            feature_mask = self.feature_mask_function(self.feature_mask)
+        # Repeat W for each batch size using broadcasting
+        W = torch.ones(x.size(0), 1, 1, device=x.device) * self.W
+        # Binning features
+        x = torch.bmm(x.unsqueeze(-1), W) - self.cut_points.unsqueeze(0)
+        x = self.binning_activation(x)  # , dim=-1)
+        x = x * self.leaf_responses.unsqueeze(0)
+        x = (x * feature_mask.reshape(1, -1, 1)).sum(dim=1)
+        return x, feature_mask
+
+
+class NeuralDecisionTree(nn.Module):
+    def __init__(
+        self,
+        depth: int,
+        n_features: int,
+        dropout: float = 0,
+        binning_activation: Callable = entmax15,
+        feature_mask_function: Callable = entmax15,
+        feature_sparsity: float = 0.8,
+        learnable_sparsity: bool = True,
+    ):
+        super().__init__()
+        self.depth = depth
+        self._num_cutpoints = 1
+        self.n_features = n_features
+        self._dropout = dropout
+        self.binning_activation = binning_activation
+        self.feature_mask_function = feature_mask_function
+        self.feature_sparsity = feature_sparsity
+        self.learnable_sparsity = learnable_sparsity
+        self._build_network()
+
+    def _build_network(self):
+        for d in range(self.depth):
+            for n in range(max(2 ** (d), 1)):
+                self.add_module(
+                    f"decision_stump_{d}_{n}",
+                    NeuralDecisionStump(
+                        self.n_features + (2 ** (d) if d > 0 else 0),
+                        self.binning_activation,
+                        self.feature_mask_function,
+                        self.feature_sparsity,
+                        self.learnable_sparsity,
+                    ),
+                )
+        self.dropout = nn.Dropout(self._dropout)
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        tree_input = x
+        feature_masks = []
+        for d in range(self.depth):
+            layer_nodes = []
+            layer_feature_masks = []
+            for n in range(max(2 ** (d), 1)):
+                leaf_nodes, feature_mask = self._modules[f"decision_stump_{d}_{n}"](
+                    tree_input
+                )
+                layer_nodes.append(leaf_nodes)
+                layer_feature_masks.append(feature_mask)
+            layer_nodes = torch.cat(layer_nodes, dim=1)
+            tree_input = torch.cat([x, layer_nodes], dim=1)
+            feature_masks.append(layer_feature_masks)
+        return self.dropout(layer_nodes), feature_masks
